requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
python-decouple==3
djangorestframework==3.11
drf-yasg
